requirements.txt
domeneshop
python-dotenv
rich